Rename folders of models
