+++ src/components/SortAndSelect.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Row, Radio, Select, Button } from 'antd';
-import { SelectorProps } from '../types/modelsOverviewTeacher';
+import { SelectorProps } from '../types/overview-teacher-models';
 
 const { Option } = Select;
