+++ src/components/charts/BarChart.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Bar } from 'react-chartjs-2';
-import { InputBarChart } from '../../types/modelsCharts';
+import { InputBarChart } from '../../types/charts-models';
 
 export default function BarChart({
   labels,
