+++ src/components/charts/DoughnutChart.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Doughnut } from 'react-chartjs-2';
-import { InputDoughnutChart } from '../../types/modelsCharts';
+import { InputDoughnutChart } from '../../types/charts-models';
 
 export default function DoughnutChart({
   data,
