+++ src/store/schoolInfo/actions.ts
@@ -2,7 +2,7 @@ import axios from 'axios';
 import { apiUrl } from '../../config/constants';
 import { Dispatch } from 'redux';
 import { STORE_TEACHERS, GetState, StoreTeachers } from './types';
-import { TeacherOption } from '../../types/school-info-model';
+import { TeacherOption } from '../../types/school-info-models';
 import { appLoading, appDoneLoading, setMessage } from '../appState/actions';
 
 export const storeTeachers = (teachers: TeacherOption[]): StoreTeachers => {
