+++ src/types/school-info-models.ts
